Convert Home Adress to Typescript
