--- playbook/app/pb_kits/playbook/pb_home_address_street/_home_address_street.jsx
@@ -1,124 +1,0 @@
-/* @flow */
-
-import React from 'react'
-import classnames from 'classnames'
-
-import { joinPresent, titleize } from '../utilities/text'
-import { globalProps } from '../utilities/globalProps'
-
-import Body from '../pb_body/_body'
-import Hashtag from '../pb_hashtag/_hashtag'
-import Title from '../pb_title/_title'
-
-type HomeAddressStreetProps = {
-  address: string,
-  addressCont: string,
-  city: string,
-  className?: string,
-  dark?: boolean,
-  emphasis: "street" | "city",
-  homeId: number,
-  houseStyle: string,
-  homeUrl: string,
-  state: string,
-  zipcode: string,
-  territory: string,
-}
-
-const HomeAddressStreet = (props: HomeAddressStreetProps) => {
-  const {
-    address,
-    addressCont,
-    city,
-    className,
-    dark = false,
-    emphasis = 'street',
-    homeId,
-    homeUrl,
-    houseStyle,
-    state,
-    zipcode,
-    territory,
-  } = props
-
-  const classes = (className, dark) =>
-    classnames(
-      {
-        'pb_home_address_street': !dark,
-        'pb_home_address_street_dark': dark,
-      },
-      globalProps(props),
-      className
-    )
-  return (
-    <div className={classes(className, dark)}>
-      {
-        <Choose>
-          <When condition={emphasis == 'street'}>
-            <div>
-              <Title
-                  className="pb_home_address_street_address"
-                  dark={dark}
-                  size={4}
-              >
-                {joinPresent([titleize(address), houseStyle], ' · ')}
-              </Title>
-              <Title
-                  className="pb_home_address_street_address"
-                  dark={dark}
-                  size={4}
-              >
-                {titleize(addressCont)}
-              </Title>
-              <Body color="light">
-                {`${titleize(city)}, ${state} ${zipcode}`}
-              </Body>
-            </div>
-          </When>
-          <When condition={emphasis == 'city'}>
-            <div>
-              <Body color="light">
-                {joinPresent([titleize(address), houseStyle], ' · ')}
-              </Body>
-              <Body color="light">{titleize(addressCont)}</Body>
-              <div>
-                <Title
-                    className="pb_home_address_street_address"
-                    dark={dark}
-                    size={4}
-                    tag="span"
-                >
-                  {`${titleize(city)}, ${state}`}
-                </Title>
-                <Body
-                    color="light"
-                    tag="span"
-                >
-                  {` ${zipcode}`}
-                </Body>
-              </div>
-            </div>
-          </When>
-        </Choose>
-      }
-
-      <If condition={homeId}>
-        <Hashtag
-            classname="home-hashtag"
-            dark={dark}
-            text={homeId}
-            type="home"
-            url={homeUrl || '#'}
-        />
-      </If>
-      <Body
-          color="light"
-          tag="span"
-      >
-        <small>{territory}</small>
-      </Body>
-    </div>
-  )
-}
-
-export default HomeAddressStreet
