+++ playbook/app/pb_kits/playbook/pb_home_address_street/docs/_home_address_street_emphasis.jsx
@@ -9,7 +9,7 @@ const HomeAddressStreetEmphasis = (props) => {
           address="70 Prospect Ave"
           addressCont="Apt M18"
           city="West Chester"
-          homeId={8250263}
+          homeId="8250263"
           homeUrl="https://powerhrg.com/"
           houseStyle="Colonial"
           state="PA"
@@ -24,7 +24,7 @@ const HomeAddressStreetEmphasis = (props) => {
           addressCont="Apt M18"
           city="West Chester"
           emphasis="city"
-          homeId={8250263}
+          homeId="8250263"
           homeUrl="https://powerhrg.com/"
           houseStyle="Colonial"
           state="PA"
